git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/branches/geant4/_symbols/09-04-ref-09_branch@52960 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/visualization/OpenGL/src/G4OpenGLXViewer.cc
@@ -345,7 +345,7 @@ void G4OpenGLXViewer::CreateFontLists () {
 void G4OpenGLXViewer::DrawText(const char * textString,double x,double y,double z, double size){
   
   // gl2ps or GL window ?
-  if (! drawGl2psText(textString,size)) {
+  if (! drawGl2psText(textString,(int)size)) {
 
     G4int font_base = G4OpenGLFontBaseStore::GetFontBase(this,(int)size);
     if (font_base < 0) {
